ajustes para a api v2

+++ src/Fluxus.Core/Models/Invoice.cs
@@ -5,7 +5,7 @@ namespace Fluxus.Core.Models;
 
 public class Invoice
 {
-    public int Id { get; private set; }
+    public long Id { get; private set; }
     public string? Description { get; private set; }
     public DateTime IssueDate { get; private set; }
     public decimal SubtotalService { get; private set; }
@@ -13,7 +13,7 @@ public class Invoice
     public decimal Total { get; private set; }
 
 
-    public Invoice(int id, string? description, DateTime issueDate, decimal subtotalService, decimal subtotalMileageAllowance, decimal total)
+    public Invoice(long id, string? description, DateTime issueDate, decimal subtotalService, decimal subtotalMileageAllowance, decimal total)
     {
         Id = id;
         Description = description;
