+++ src/Fluxus.Core/Models/Service.cs
@@ -4,14 +4,14 @@ namespace Fluxus.Core.Models;
 
 public class Service
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Tag { get; set; }
     public string? Description { get; set; }
     public decimal ServiceAmount { get; set; }
     public decimal MileageAllowance { get; set; }
 
 
-    public Service(int id, string tag, string description, decimal serviceAmount, decimal mileageAllowance)
+    public Service(long id, string tag, string description, decimal serviceAmount, decimal mileageAllowance)
     {
         Id = id;
         Tag = tag;
